==> r_scripts/1_package_info.Rmd <==
---
title: "1_package_info"
author: "Hailey Sounart"
date: "10/20/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(include = TRUE, echo = TRUE, message = TRUE, warning =FALSE)
```

## Load all packages used in analysis

```{r load packages}
  library(STutility)
  library(stringr)
  library(openxlsx)
  library(tidyr)
  library(ggplot2)
  library(ggpubr)
  library(dplyr)
  library(harmony)
  library(EnhancedVolcano)
  library(magrittr)
  library(magick)
  library(grid)
```

## Get session information

```{r pressure, echo=FALSE}

sessioninfo::session_info()

```
Output is the versions of all package dependencies for running the R analysis scripts 2_make_brugia_wolbachia_gene_lists.Rmd, 3_add_metadata.Rmd, 4_quality_check_filtering.Rmd, 5_normalization.Rmd, 6_clustering_analysis.Rmd, 7_co_localization_analysis.Rmd, 8_genes_on_tissue_visualization.Rmd, and 9_wolbachia_abundance_variation.Rmd.
