==> r_scripts/8_genes_on_tissue_visualization.Rmd <==
---
title: "8_genes_on_tissue_visualization"
author: "Hailey Sounart"
date: "9/13/2023"
output: html_document
---

```{r, include = TRUE, echo = TRUE, message = TRUE}

start_time <- Sys.time()
start_time
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(include = TRUE, echo = TRUE, message = FALSE, warning =FALSE)
```

```{r load packages}
  library(STutility)
  library(magrittr)
  library(ggplot2)
  library(magick)
  library(grid)
```

```{r set working directory}

setwd("<path to working directory where count matrices for all samples reside>")

``` 

# Read in clustered object

``` {r load clustered object}

brugia_clu_obj <- readRDS("bmalayi_clusters.rds")
brugia_clu_obj

```

```{r rename clus numbers}

new_clus_names <- brugia_clu_obj
new_clus_names$seurat_clusters <- factor(brugia_clu_obj$seurat_clusters, 
         levels = c("0", "1", "2", "3"), 
         labels = c("1", "2", "3", "4"))

new_clus_names@meta.data
```

```{r make wBm metadata}

#add wolbachia column to metadata
new_clus_names@meta.data$wolbachia <- ifelse((new_clus_names@meta.data$counts.wBm >= 1), 1, 0)
new_clus_names@meta.data

```

# Genes to visualize 

```{r load genes gly}

# Figure 4 genes

gly_genes <- c("WBGene00225841","WBGene00225960","WBGene00223600")
gly_genes

gly_genes_rename <- paste0("bmalayi-", gly_genes)
gly_genes_rename

```

```{r colocal genes}

# Figure 5 genes

colocal_genes <- c("WBGene00222232","WBGene00224389")
colocal_genes

colocal_genes_rename <- paste0("bmalayi-", colocal_genes)
colocal_genes_rename

```

# Load raw images
```{r load raw image}

# Path to the raw HE images used as input for spaceranger

high_res_image_list <- c("/spaceranger-inputs/V10F24_041_A1/V10F24_041_A1_small.jpg",
                         "/spaceranger-inputs/V10F24_041_B1/V10F24_041_B1_small.jpg",
                         "/spaceranger-inputs/V10F24_041_D1/V10F24_041_D1_small.jpg")
```

# Load the images
```{r load images raw,unfilt, fig.width = 10, fig.height = 10}

bmalayi_obj <- LoadImages(new_clus_names, time.resolve = F, verbose = T)

``` 

# Get image pixels
```{r get image pixels}

# Now we need the pixel coordinates to make it possible to draw points on the capture area
# Most image related info is stored in the "Staffli" object which can be accessed using the
# GetStaffli function

bmalayi_obj.object <- GetStaffli(bmalayi_obj)
bmalayi_obj.object

bmalayi_obj.object@dims[[1]]$width
bmalayi_obj.object@dims[[1]]$height

```

# Set Resolution

```{r set resolution}

# Now we can load the original HE image in whatever resolution
# Here, it's loaded to a width of 5000 pixels with a fixed aspect ratio
# set a size that is smaller than the original image.

imwidth <- 5000
section <- c(1:3)

```

# Get Image Info

```{r get image info}

# Load the image and retrieve some info about it. Takes some time if the image was
# exported in full resolution from the microscope. Do NOT plot full res image.

im_list <- lapply(1:length(high_res_image_list), function(i) {image_read(high_res_image_list[[i]])})
im_list

```

# Scale Down Image

```{r scale down image }

# An image of 5000 pixels in width is big to draw in the RStudio Viewer
# Scale down the original image to our predefined imwidth and convert it to raster format.

im_list <- lapply(1:length(im_list), function(i) {image_scale(im_list[[i]], paste0(imwidth)) %>% as.raster()})

lapply(1:length(im_list), function(i) {
  print(class(im_list[[i]]))
  print(dim(im_list[[i]]))})

```

# Scale Pixel Coordinates

```{r scale pixel coordinates}

# Next, we need to scale our pixel coordinates to fit our new image. The pixel coordinates
# are stored in the st.object meta.data slot. If  multiple samples are present here,
# first subset this meta.data to include the section of interest.

gg_list <- lapply(1:length(im_list), function(i) {subset(bmalayi_obj.object@meta.data, sample %in% paste0(section[i]))[, c("pixel_x", "pixel_y")]}) 
gg_list 

```

# Get Scaling Factor

```{r get scaling factor}

# We can get the scaling factor by dividing the width of our new image by the
# width of our images that were used when running InputFromTable, i.e. the
# "tissue_hires_image.png".

bmalayi_obj.object@dims[[1]]$width

scale.factor_list <- lapply(1:length(im_list), function(i) {dim(im_list[[i]])[2]/bmalayi_obj.object@dims[[1]]$width})

scale.factor_list

```

# Get New Coordinates

```{r get new coordinates}

# Then we multiply our pixel coordinates by this scale.factor to get our new coordinates
 
gg_list <- Map('*',gg_list,scale.factor_list)
gg_list 

```


# Add Feature Values - clusters

```{r add features values}

# Next, add the feature values you want to draw.
# We can use the FetchData function which collects values from different slots

data_list_clu <- lapply(1:length(gg_list), function(i) {FetchData(object = bmalayi_obj, vars = "seurat_clusters", cells = rownames(subset(bmalayi_obj.object@meta.data, sample %in% paste0(section[i]))))})

data_list_clu

gg_list_clu <- Map(cbind , gg_list, data_list_clu)

gg_list_clu 

```

# Visualize clusters on tissue
```{r ggplot magic seurat clusters}

# Figure 2C

# Most important is to get the color scaling to be correct. If plotting a centered vector (centered at 0), make sure
# to center the colorscale as well, with the scale_color_gradientn() function. If plotting a feature with strictly 
# positive values, get rid of the limits option. If plotting a categorical variable, like clusters, control the colors 
# using scale_color_manual() instead of scale_color_gradientn().

fill=c("#41ab5d","#fc9272","#4292c6","#dd648f")

selected.feature <- "seurat_clusters"
g <- lapply(1:length(section), function(i) grid::rasterGrob(im_list[[i]], width = unit(1, "npc"), height = unit(1, "npc"), interpolate = TRUE))
for(i in seq_along(gg_list_clu )) {
  g_sub = g[[i]]
  df = as.data.frame(gg_list_clu [i])
  colScale <- scale_colour_manual(name = "seurat_clusters",values = fill, drop=F)
  
  p <- ggplot(df, aes(color=selected.feature)) +
    annotation_custom(g_sub, -Inf, Inf, -Inf, Inf) +
    geom_point(aes_string("pixel_x", "dim(im_list[[i]])[1] - pixel_y", color=selected.feature), size = 6.5) + 
    scale_x_continuous(limits = c(0, dim(im_list[[i]])[2]), expand = c(0, 0)) +
    scale_y_continuous(limits = c(0, dim(im_list[[i]])[1]), expand = c(0, 0)) +
    theme_void() 
  p1 <- p + colScale
    pdf(file = paste0("Figure_2C_",  unique(bmalayi_obj$slide_subarray)[i] ,".pdf"), width = 33.336, height = 32.28)
  print(p1)
  dev.off()
}

```

# Add Feature Values - genes

```{r add gene values}

ftrs <- c(gly_genes_rename, colocal_genes_rename, "wolbachia")
ftrs 

data_list <- lapply(1:length(gg_list), function(i) {FetchData(object = bmalayi_obj, vars = ftrs, cells = rownames(subset(bmalayi_obj.object@meta.data, sample %in% paste0(section[i]))))}) 

data_list

gg_list <- Map(cbind , gg_list, data_list)

gg_list 

```

``` {r make dir}

dir.create("genes_on_tissue")

for (gene in ftrs) {
  dir.create(paste0("genes_on_tissue/",gene))
}

ftrs.edited <- gsub("-", ".", ftrs)
ftrs.edited

gly_edited <- ftrs.edited[1:3]
gly_edited
colocal_edited <- ftrs.edited[4:5]
colocal_edited

```

# Visualize glycolytic genes on tissue
```{r plot gly genes}

# Figure 4B-D

for (feat in gly_edited) {
  selected.feature <- feat
  folder.name <- gsub("bmalayi.", "bmalayi-", feat)
  
  g <- lapply(1:length(section), function(i) grid::rasterGrob(im_list[[i]], width = unit(1, "npc"), height = unit(1, "npc"), interpolate = TRUE))
  
  for(i in seq_along(gg_list)) {
    g_sub = g[[i]]
    df = as.data.frame(gg_list[i])
    
    p <- ggplot() +
    annotation_custom(g_sub, -Inf, Inf, -Inf, Inf) +
    geom_point(data = df, aes_string("pixel_x", "dim(im_list[[i]])[1] - pixel_y", color = selected.feature), size = 8) +
    scale_x_continuous(limits = c(0, dim(im_list[[i]])[2]), expand = c(0, 0)) +
    scale_y_continuous(limits = c(0, dim(im_list[[i]])[1]), expand = c(0, 0)) +
    theme_void() +
    scale_color_gradientn(colors = c("navyblue", "cyan", "yellow", "red", "dark red"), limits = c(0, max(abs(df[, selected.feature])))) +
    theme(legend.position = c(0.89, 0.85), legend.key.height = unit(x = 1.2, units = "cm"),
          legend.key.width = unit(x = 1.2, units = "cm"),
          legend.text = element_text(size = 20),
          legend.title = element_text(size = 20))
      theme_void() 

    pdf(file = paste0("genes_on_tissue/", folder.name, "/", "Figure_4_", folder.name, "_", unique(bmalayi_obj$slide_subarray)[i] ,".pdf"), width = 33.336, height = 32.28)
  print(p)
  dev.off()
  }
}

```

# Visualize colocal genes on tissue - W+ spots
```{r plot colocal wol pos}

# Figure 5C-D

for (feat in colocal_edited) {
  selected.feature <- feat
  folder.name <- gsub("bmalayi.", "bmalayi-", feat)
  
  g <- lapply(1:length(section), function(i) grid::rasterGrob(im_list[[i]], width = unit(1, "npc"), height = unit(1, "npc"), interpolate = TRUE))
  
  for(i in seq_along(gg_list)) {
    g_sub = g[[i]]
    df = as.data.frame(gg_list[i])
    df_wol <- df[(df$wolbachia == "1"),]
    
    p <- ggplot() +
    annotation_custom(g_sub, -Inf, Inf, -Inf, Inf) +
    geom_point(data = df_wol, aes_string("pixel_x", "dim(im_list[[i]])[1] - pixel_y", color = selected.feature), size = 8) +
    scale_x_continuous(limits = c(0, dim(im_list[[i]])[2]), expand = c(0, 0)) +
    scale_y_continuous(limits = c(0, dim(im_list[[i]])[1]), expand = c(0, 0)) +
    theme_void() +
    scale_color_gradientn(colors = c("navyblue", "cyan", "yellow", "red", "dark red"), limits = c(0, max(abs(df[, selected.feature])))) +
    theme(legend.position = c(0.89, 0.85), legend.key.height = unit(x = 1.2, units = "cm"),
          legend.key.width = unit(x = 1.2, units = "cm"),
          legend.text = element_text(size = 20),
          legend.title = element_text(size = 20))
      theme_void() 

    pdf(file = paste0("genes_on_tissue/", folder.name, "/", "Figure_5_", folder.name, "_", unique(bmalayi_obj$slide_subarray)[i],"_", "wol_1",".pdf"), width = 33.336, height = 32.28)
  print(p)
  dev.off()
  }
}

```

# Visualize colocal genes on tissue - W- spots
```{r plot colocal wol neg}

# Figure 5C-D

for (feat in colocal_edited) {
  selected.feature <- feat
  folder.name <- gsub("bmalayi.", "bmalayi-", feat)
  
  g <- lapply(1:length(section), function(i) grid::rasterGrob(im_list[[i]], width = unit(1, "npc"), height = unit(1, "npc"), interpolate = TRUE))
  
  for(i in seq_along(gg_list)) {
    g_sub = g[[i]]
    df = as.data.frame(gg_list[i])
    df_no_wol <- df[(df$wolbachia == "0"),]
    
    p <- ggplot() +
    annotation_custom(g_sub, -Inf, Inf, -Inf, Inf) +
    geom_point(data = df_no_wol, aes_string("pixel_x", "dim(im_list[[i]])[1] - pixel_y", color = selected.feature), size = 8) +
    scale_x_continuous(limits = c(0, dim(im_list[[i]])[2]), expand = c(0, 0)) +
    scale_y_continuous(limits = c(0, dim(im_list[[i]])[1]), expand = c(0, 0)) +
    theme_void() +
    scale_color_gradientn(colors = c("navyblue", "cyan", "yellow", "red", "dark red"), limits = c(0, max(abs(df[, selected.feature])))) +
    theme(legend.position = c(0.89, 0.85), legend.key.height = unit(x = 1.2, units = "cm"),
          legend.key.width = unit(x = 1.2, units = "cm"),
          legend.text = element_text(size = 20),
          legend.title = element_text(size = 20))
      theme_void() 

    pdf(file = paste0("genes_on_tissue/", folder.name, "/", "Figure_5_", folder.name, "_", unique(bmalayi_obj$slide_subarray)[i],"_", "no_wol_0",".pdf"), width = 33.336, height = 32.28)
  print(p)
  dev.off()
  }
}

```

# Wolbachia spots binary on tissue
```{r plot wol binary}

# Figure 5C-D

selected.feature <- "wolbachia"
g <- lapply(1:length(section), function(i) grid::rasterGrob(im_list[[i]], width = unit(1, "npc"), height = unit(1, "npc"), interpolate = TRUE))

for(i in seq_along(gg_list)) {
  g_sub = g[[i]]

  p <- ggplot() +
  annotation_custom(g_sub, -Inf, Inf, -Inf, Inf) +
  geom_point(data = gg_list[[i]], aes_string("pixel_x", "dim(im_list[[i]])[1] - pixel_y", color = as.factor(gg_list[[i]]$wolbachia)), size = 8) +
  scale_x_continuous(limits = c(0, dim(im_list[[i]])[2]), expand = c(0, 0)) +
  scale_y_continuous(limits = c(0, dim(im_list[[i]])[1]), expand = c(0, 0)) +
  theme_void() +
  scale_color_manual(values = c("#00FF7F","#FA8072")) +
  theme(legend.position = c(0.89, 0.85), legend.key.height = unit(x = 1.2, units = "cm"),
        legend.key.width = unit(x = 1.2, units = "cm"),
        legend.text = element_text(size = 20),
        legend.title = element_text(size = 20))
    theme_void() 
  
  pdf(file = paste0("genes_on_tissue/wolbachia/Figure_5_binary_colocal_", unique(bmalayi_obj$sample_section)[i] ,".pdf"), width = 33.336, height = 32.28)
print(p)
dev.off()
}

```

```{r, include = TRUE, echo = TRUE, message = TRUE}

end_time <- Sys.time()
end_time
```

```{r, include = TRUE, echo = TRUE, message = TRUE}

elapsed_time <- round((end_time-start_time), 3)
elapsed_time
```

Output is clusters plotted on tissue sections, shown in Figure 2C, and genes plotted on tissue sections, shown in Figure 4B-D and Figure 5C-D.
